+++ Reinforcement/Updaters/RegisterUpdater.cs
@@ -33,10 +33,10 @@ namespace Updaters
                 var updaterId = changedWidthUpdater.GetUpdaterId();
 
                 var filter = new ElementCategoryFilter(BuiltInCategory.OST_GenericAnnotation);
-                UpdaterRegistry.AddTrigger(updaterId, filter, Element.GetChangeTypeParameter(new ElementId(590069)));
-                UpdaterRegistry.AddTrigger(updaterId, filter, Element.GetChangeTypeParameter(new ElementId(590070)));
-                UpdaterRegistry.AddTrigger(updaterId, filter, Element.GetChangeTypeParameter(new ElementId(10585865)));
-                UpdaterRegistry.AddTrigger(updaterId, filter, Element.GetChangeTypeParameter(new ElementId(10585866)));
+                UpdaterRegistry.AddTrigger(updaterId, filter, Element.GetChangeTypeParameter(new ElementId((long)590069)));
+                UpdaterRegistry.AddTrigger(updaterId, filter, Element.GetChangeTypeParameter(new ElementId((long)590070)));
+                UpdaterRegistry.AddTrigger(updaterId, filter, Element.GetChangeTypeParameter(new ElementId((long)10585865)));
+                UpdaterRegistry.AddTrigger(updaterId, filter, Element.GetChangeTypeParameter(new ElementId((long)10585866)));
             }
             else
             {
